"CheckBoxRequired moved to folder AuthHelpers"
