+++ RikaWebApp/Helpers/AuthHelpers/CheckBoxRequired.cs
@@ -1,6 +1,6 @@
 using System.ComponentModel.DataAnnotations;
 
-namespace RikaWebApp.Helpers;
+namespace RikaWebApp.Helpers.AuthHelpers;
 public class CheckBoxRequired : ValidationAttribute
 {
     public override bool IsValid(object? value) => value is bool b && b;
